pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dartfx-mtnards"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [
  { name = "kulnor", email = "pascal.heus@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "markdownify>=0.13.0",
  "mcp[cli]>=1.6.0",
  "mlcroissant>=1.0.16",
  "pydantic>=2.0.0",
  "requests>=2.0.0",
  "python-dotenv>=1.0.0",
  "rdflib>=7.0.0",
]

[project.urls]
Documentation = "https://github.com/DataArtifex/mtnards-toolkit#readme"
Issues = "https://github.com/DataArtifex/mtnards-toolkit/issues"
Source = "https://github.com/DataArtifex/mtnards-toolkit"


[project.optional-dependencies]
mcp = [
  "httpx>=0.20.0",
  "mcp>=1.6",
  "mcp[cli]>=1.6",
]

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
  "/docs",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dartfx"]

[tool.hatch.version]
path = "src/dartfx/mtnards/__about__.py"

[dependency-groups]
dev = [
  "coverage[toml]>=6.5",
  "pytest>=7.0.0",
  "pytest-benchmark>=4.0.0",
  "mypy>=1.0.0",
  "python-dotenv>=1.0.0",
  "ruff>=0.3.0",
  "pre-commit>=3.0.0",
]
docs = [
  "sphinx>=7.0.0",
  "sphinx-rtd-theme>=1.2.0",
  "myst-parser>=1.0.0",
  "sphinx-copybutton>=0.5.0",
]

[tool.hatch.envs.default]
dependencies = ["dev"]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.12", "3.13"]

[tool.hatch.envs.types]
dependencies = ["dev"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/dartfx/mtnards tests}"

[tool.hatch.envs.docs]
dependencies = ["docs"]
[tool.hatch.envs.docs.scripts]
build = "sphinx-build -b html docs/source docs/build/html"
serve = "python -m http.server -d docs/build/html"

[tool.hatch.metadata]
allow-direct-references = true

[tool.coverage.run]
source_pkgs = ["mtnards", "tests"]
branch = true
parallel = true
omit = [
  "src/dartfx/mtnards/__about__.py",
]

[tool.coverage.paths]
source = ["src/dartfx/mtnards", "*/mtnards-toolkit/src/dartfx/mtnards"]
tests = ["tests", "*/dartfx/mtnards/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "ARG", # flake8-unused-arguments
  "PT",  # flake8-pytest-style
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["dartfx"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
